requirements.txt
flask
flask-login
werkzeug
python-docx
matplotlib
